Proof read TAGS_FOR_RELATIVE_SETLOC (#357)
There was not much to fix in this part of the code except for a few comment fixes. But I am unsure as to whether the [subroutine calls on page 36](http://www.ibiblio.org/apollo/ScansForConversion/Comanche055/0036.jpg) were left out intentionally or are just missing. Should they be added?

EDIT: Also, this closes #263